requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
pydantic==2.5.2
pydantic-settings==2.1.0